Page service refactoring
