+++ TheraLang.Web/ClientApp/src/app/modules/cms-generic/components/piranha-page/piranha-page.component.ts
@@ -1,6 +1,6 @@
 import { Page } from "../../../../shared/models/page/page.model";
 import { Component, OnInit, Input } from "@angular/core";
-import { CmsPageService } from "../../../../core/http/cms/cms-page.service";
+import { PageService } from "../../../../core/http/manager/page.service";
 import { CmsRoute } from "src/app/modules/main/toolbar/toolbar-item/cms-route";
 
 @Component({
@@ -12,12 +12,12 @@ export class PiranhaPageComponent implements OnInit {
   page: Page;
   @Input() cmsRoute: CmsRoute;
   ifGenerate = false;
-  constructor(private cmsPageService: CmsPageService) {}
+  constructor(private cmsPageService: PageService) {}
 
   async ngOnInit() {
-    this.page = await this.cmsPageService.getPiranhaPageById(
-      parseInt(this.cmsRoute.id)
-    );
+    // this.page = await this.cmsPageService.getPageById(
+    //   parseInt(this.cmsRoute.id)
+    // );
     this.ifGenerate = true;
   }
 }
