+++ TheraLang.Web/ClientApp/src/app/modules/manager/page-manager/edit-page/edit-page.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FormGroup, Validators, FormControl } from '@angular/forms';
-import { Newpage } from 'src/app/shared/models/new_page/newpage';
-import { PageService } from '../../shared/services/page.service';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { switchMap } from 'rxjs/operators';
+import { Page } from 'src/app/shared/models/page/page.model';
+import { PageService } from 'src/app/core/http/manager/page.service';
+import { NotificationService } from 'src/app/core/services/notification/notification.service';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-edit-page',
@@ -11,7 +13,7 @@ import { switchMap } from 'rxjs/operators';
   styleUrls: ['./edit-page.component.less']
 })
 export class EditPageComponent implements OnInit {
-  page: Newpage;
+  page: Page;
   form = new FormGroup({
     header: new FormControl('', [Validators.required, Validators.maxLength(60)]),
     menuName: new FormControl('', [Validators.required, Validators.maxLength(60)]),
@@ -21,7 +23,9 @@ export class EditPageComponent implements OnInit {
   constructor(
     private pageService: PageService,
     private router: Router,
-    private route: ActivatedRoute
+    private route: ActivatedRoute,
+    private notificationService: NotificationService,
+    private translate: TranslateService
   ) {
   }
 
@@ -30,7 +34,7 @@ export class EditPageComponent implements OnInit {
       switchMap((params: Params) => {
         return this.pageService.getPageById(params['id']);
       })
-    ).subscribe((page: Newpage) => {
+    ).subscribe((page: Page) => {
       this.page = page
       this.form.controls.header.setValue(page.header);
       this.form.controls.content.setValue(page.content);
@@ -56,6 +60,18 @@ export class EditPageComponent implements OnInit {
       header: this.form.value.header,
       menuName: this.form.value.menuName,
       content: this.form.value.content
-    });
+    }).subscribe(
+      async (msg: string) => {
+        msg = await this.translate.get("common.updated-successfully").toPromise();
+        this.notificationService.success(msg);
+        this.router.navigate(['admin', 'sitemap']);
+      },
+      async error => {
+        console.log(error);
+        this.notificationService.warn(
+          await this.translate.get("common.wth").toPromise()
+        );
+      }
+    );
   }
 }
