--- TheraLang.Web/ClientApp/src/app/shared/models/new_page/newpage.ts
@@ -1,6 +1,0 @@
-export class Newpage {
-    id?: number;
-    header: string;
-    content: string;
-    menuName: string;
-}
