+++ TheraLang.Web/ClientApp/src/app/shared/models/page/page.model.ts
@@ -1,28 +1,6 @@
-import { Block } from "../block/block.model";
-import { TopTitle } from "../top-title/top-title.model";
-
 export class Page {
-  hero: TopTitle;
-  isStartPage: boolean;
-  siteId: string;
-  contentType: string;
-  parentId: string;
-  sortOrder: number;
-  navigationTitle?: any;
-  isHidden: boolean;
-  redirectUrl?: any;
-  redirectType: number;
-  originalPageId?: any;
-  blocks: Block[];
-  slug: string;
-  permalink: string;
-  metaKeywords?: any;
-  metaDescription?: any;
-  route: string;
-  published: string;
-  id: string;
-  typeId: string;
-  title: string;
-  created: string;
-  lastModified: string;
+  id?: number;
+  header: string;
+  content: string;
+  menuName: string;
 }
